発注ロジックをシンプル版に作り直す：定数・在庫・減り数・切れる日・入庫で延命の1画面
利用者の4点（在庫CSVを貼る／動員で減り数が決まる／「○/○ に切れる」のアラート／
入庫日と入庫数で延命）だけに絞り、発注点・安全在庫・保管スペース・発注サイクル・
消費量推定・ダッシュボードを外した。以前の版は excel/legacy/ に退避。

- ②発注する：定数(個)・在庫(個)・減り数/日・切れる日・「このままだと」・入庫日/入庫数
  （2便目まで）・目安・「入庫を入れると」・14日の在庫表。上部に ● △ ⚠ の件数と発注予定額
- 在庫の予測は MAX(0, 前日＋入庫−減り) で0に止める（切れた後の架空の赤字を作らない）
- 行の並びは商品マスタの順に固定（CSVの増減で入庫の入力が別商品にずれない）。
  マスタに無い商品は末尾に出し、ホームで追加を促す
- 誤操作対策：入力セル以外はシート保護、入庫日は表の14日から選ぶドロップダウン、
  入庫数は目安（定数まで）を上限、過去日・片方だけ・定数超え・減り数未設定・動員未入力・
  古いCSV・同じ商品コードの重複行・動員の表の範囲外を文字で警告。⚠の行は発注書に載せない
- 定数はマスタの定数（ケース）×入数、無ければ対象劇場のCSV規定数。CSVの数値は対象劇場の行だけから取る
- 動員の表は開始日（空なら今年の1/1）から400日分。曜日の実績平均は直近8週
- 納品する xlsx は openpyxl の出力そのまま（LibreOffice で再計算保存するとフォントとリンクが変わる）。
  Excel が開くときに全計算する
- verify.py：Python で同じ計算をやり直して全行を突き合わせる（サンプルで一致 0件差）
- make_sample.py：3劇場CSV・定数・減り数・動員・入庫の例（目安どおり／少ない／遅い／定数超え／
  過去日／2便目／数えた在庫）を入れる。再計算は控えのコピーで行う
- README を全面改訂
